지점 게시판 첨부 유지

+++ src/main/java/com/unvus/firo/module/adapter/azure/AzureAdapter.java
@@ -61,19 +61,19 @@ public class AzureAdapter implements Adapter {
     }
 
     @Override
-    public void writeFromTemp(DirectoryPathPolicy directoryPathPolicy, String fullDir, String path, String tempFileName, long size, String contentType) throws Exception {
+    public void writeFromTemp(DirectoryPathPolicy directoryPathPolicy, String fullDir, String path, String tempFileName, long size, String contentType, boolean keepTemp) throws Exception {
         String fullPath = Adapter.adjustSeparator(Paths.get(fullDir, path), directoryPathPolicy.getSeparator());
         String tempFullPath = Adapter.adjustSeparator(Paths.get(directoryPathPolicy.getTempDir(), tempFileName), directoryPathPolicy.getSeparator());
 
-        rename(tempFullPath, fullPath);
+        rename(tempFullPath, fullPath, keepTemp);
     }
 
     @Override
-    public void rename(String from, String to) throws Exception {
+    public void rename(String from, String to, boolean keepFrom) throws Exception {
         BlobClient sourceBlob = containerClient.getBlobClient(adjustPath(from));
         BlobClient destBlob = containerClient.getBlobClient(adjustPath(to));
         destBlob.copyFromUrl(sourceBlob.getBlobUrl());
-        if (destBlob.exists()) {
+        if (destBlob.exists() && !keepFrom) {
             sourceBlob.delete();
         }
     }
